move some commands to new server cog, lay foundation for fakeperms
