+++ utils/Checks.py
@@ -7,10 +7,12 @@ import json
 
 from utils import DBClient, CONSTANTS, CachedDB
 
+from discord.ext import commands
+from discord.ext.commands import Context
+
 db = DBClient.db
 
-async def is_not_blacklisted(context):
-
+async def is_not_blacklisted(context: Context):
     users_global = db["users_global"]
     user = await CachedDB.find_one(users_global, {"id": context.author.id}, ex=120)
 
@@ -25,5 +27,13 @@ async def is_not_blacklisted(context):
 
 # TODO: Make this
 # Coming soon
-async def has_perm(*argv):
-    pass
+def has_perm(**perms):
+    def predicate(context: commands.Context):
+        author_permissions = context.channel.permissions_for(context.author)
+
+        for perm, value in perms.items():
+            if getattr(author_permissions, perm, None) != value:
+                raise discord.ext.commands.MissingPermissions([perm])
+        return True
+
+    return commands.check(predicate)
